requirements.txt
aloe==0.1.4
ansible==2.0.2.0
appnope==0.1.0
argh==0.26.1
astroid==1.4.4
beautifulsoup4==4.4.1
blessings==1.6
boto==2.40.0
Chameleon==2.24
colorama==0.3.7
decorator==4.0.9
dj-database-url==0.4.1
Django==1.9.5
django-registration==2.0.4
djangorestframework==3.3.3
docopt==0.6.2
easy-thumbnails==2.3
ecdsa==0.13
factory-boy==2.6.1
fake-factory==0.5.7
flake8==2.5.4
flake8-docstrings==0.2.5
future==0.15.2
gevent==1.1.0
gherkin-official==4.0.0
gnureadline==6.3.3
greenlet==0.4.9
ipaddress==1.0.16
ipython==4.1.1
ipython-genutils==0.1.0
Jinja2==2.8
lazy-object-proxy==1.2.1
Mako==1.0.4
Markdown==2.6.6
MarkupSafe==0.23
mccabe==0.4.0
nose==1.3.7
paramiko==1.16.0
passlib==1.6.5
PasteDeploy==1.5.2
path.py==8.1.2
pathlib==1.0.1
pathtools==0.1.2
pep257==0.7.0
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.6
Pillow==3.2.0
pluggy==0.3.1
postgres==2.2.1
psutil==4.1.0
psycopg2==2.6.1
ptyprocess==0.5.1
py==1.4.31
pycrypto==2.6.1
pyflakes==1.0.0
Pygments==2.1.3
pylint==1.5.4
pyramid==1.6.1
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.4.2
pyramid-ipython==0.1
pyramid-jinja2==2.6.2
pyramid-mako==1.0.2
pyramid-tm==0.12.1
pytest==2.9.0
pytest-watch==3.10.0
python-dateutil==2.5.2
PyYAML==3.11
repoze.lru==0.6
requests==2.9.1
requirements==0.1
simplegeneric==0.8.1
six==1.10.0
sorl-thumbnail==12.3
SQLAlchemy==1.0.12
thefuck==3.8
tox==2.3.1
traitlets==4.1.0
transaction==1.4.4
translationstring==1.3
venusian==1.0
virtualenv==14.0.6
waitress==0.8.10
watchdog==0.8.3
WebOb==1.5.1
WebTest==2.0.20
wrapt==1.10.6
WTForms==2.1
zope.deprecation==4.1.2
zope.interface==4.1.3
zope.sqlalchemy==0.7.6
